Fix nqp::pow_i some more.

+++ src/vm/jvm/runtime/org/perl6/nqp/runtime/Ops.java
@@ -6192,7 +6192,7 @@ public final class Ops {
     }
     
     public static long pow_i(long a, long b) {
-        return (int)Math.pow(a, b);
+        return (long)Math.pow(a, b);
     }
 
     public static double pow_n(double a, double b) {
